add createtree, leafsimilar beat 19%

+++ library/Graph.cpp
@@ -58,7 +58,7 @@ class ConnectedComponents {
 		if (visited[i])
 			return;
 		visited[i] = true;
-		for (int f = 0; f < M.size(); f++) {  // adjacency matrix define directed friendship
+		for (size_t f = 0; f < M.size(); f++) {  // adjacency matrix define directed friendship
 			if (M[i][f])
 				dfs(M, f);
 		}
@@ -68,7 +68,7 @@ public:
 	int findCircleNum(vector<vector<int>>& M) {// connected components, beat 98%
 		visited.resize(M.size());
 		int connected = 0;
-		for (int f = 0; f < M.size(); f++) {
+		for (size_t f = 0; f < M.size(); f++) {
 			if (!visited[f]) {
 				connected++;
 				dfs(M, f);
